+++ client/routes.js
@@ -4,7 +4,7 @@ import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
 import {Login, Signup, UserHome} from './components'
 import {me} from './store'
-import ProductDetail from './components/Product/ProductDetail.jsx';
+import ProductDetail from './components/Product/ProductDetail'
 import {fetchProducts} from './store/products'
 
 /**
@@ -23,12 +23,12 @@ class Routes extends Component {
         {/* Routes placed here are available to all visitors */}
         <Route path="/login" component={Login} />
         <Route path="/signup" component={Signup} />
+        <Route path="/products/:productId" component={ProductDetail} />
         {
           isLoggedIn &&
             <Switch>
               {/* Routes placed here are only available after logging in */}
               <Route path="/home" component={UserHome} />
-              <Route path='/products/:productId' component={ProductDetail} />
             </Switch>
         }
         {/* Displays our Login component as a fallback */}
