+++ server/db/models/product.js
@@ -25,7 +25,7 @@ const Product = db.define('product', {
         }
     },
     price: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(10,2),
         allowNull: false,
         validate: {
             notEmpty: true,
